email-marketing/pyproject.toml
[tool.poetry]
name = "email-marketing"
version = "0.1.0"
description = ""
authors = ["Arthur Alvim <afmalvim@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
python-decouple = "^3.5"
Jinja2 = "^3.0.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
